+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2022-07-06
+
+- Made merge-condition optional in wz pr restrictions from_map/1 conversion. 
+- Added test data for merge-condition for wz pr restrictions data generation.
+- Added delay parameter to throttle trying each repo_config
+
 ## [1.0.0] - 2020-08-16
 
 - Open sourcing of an existing internal project